smaartpulse-api/requirements.txt
beautifulsoup4==4.5.3
boto==2.45.0
boto3==1.4.8
botocore==1.8.8
Cython==0.25.2
cytoolz==0.8.2
Flask==0.12
Flask-Cors==3.0.2
gapic-google-cloud-language-v1beta2==0.15.3
gapic-google-cloud-pubsub-v1==0.15.4
gensim==3.1.0
gevent==1.2.1
google-auth==1.2.1
google-gax==0.15.16
googleapis-common-protos==1.5.3
ipython==5.1.0
matplotlib==2.0.0
nltk==3.2.2
nose==1.3.7
numpy==1.13.1
numpydoc==0.6.0
oauth2client==3.0.0
odo==0.5.0
pandas==0.21.1
Pillow==4.0.0
proto-google-cloud-language-v1beta2==0.15.3
proto-google-cloud-pubsub-v1==0.15.4
protobuf==3.5.1
psutil==5.0.1
Pygments==2.1.3
PyYAML==3.12
QtPy==1.2.1
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
smart-open==1.5.5
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
SQLAlchemy==1.1.5
tornado==4.4.2
Tree==0.2.2
unicodecsv==0.14.1
urllib3==1.21.1
webencodings==0.5.1
Werkzeug==0.11.15
statistics
MySQL-python